Changed trend params
Changed trend params to explore whether this was dropping the pixels. It seems like some of the pixels have few obs than others, which is resulting in trends not being calculated for those pixels. This makes no sense, when all the pixels should have the same number of obs due to the manual filtering. Also, whatever I've done to the package now means there is extreme greening happening, when before there was only browning.