discovered issue with 17 NA pixels on 2nd July
There is something happening where 17 pixels in the NE of the plot do not have snow cover values for the 2nd of July only. However, the spatial extent of the data should be identical in all of the rasters used to derive the snow cover. Need to investigate what is going on here.